# Optimize `BigInt.new(Int::Primitive)`