--- a/src/big/lib_gmp.cr
+++ b/src/big/lib_gmp.cr
@@ -29,18 +29,21 @@ lib LibGMP
 
   {% if flag?(:win32) && flag?(:bits64) %}
     alias MpExp = LibC::Long
+    alias MpSize = LibC::LongLong
     alias MpLimb = LibC::ULongLong
   {% elsif flag?(:bits64) %}
     alias MpExp = Int64
+    alias MpSize = LibC::Long
     alias MpLimb = LibC::ULong
   {% else %}
     alias MpExp = Int32
+    alias MpSize = LibC::Long
     alias MpLimb = LibC::ULong
   {% end %}
 
   struct MPZ
-    _mp_alloc : Int32
-    _mp_size : Int32
+    _mp_alloc : Int
+    _mp_size : Int
     _mp_d : MpLimb*
   end
 
@@ -143,6 +146,11 @@ lib LibGMP
   fun lcm = __gmpz_lcm(rop : MPZ*, op1 : MPZ*, op2 : MPZ*)
   fun lcm_ui = __gmpz_lcm_ui(rop : MPZ*, op1 : MPZ*, op2 : UI)
   fun remove = __gmpz_remove(rop : MPZ*, op : MPZ*, f : MPZ*) : BitcntT
+
+  # Special Functions
+
+  fun limbs_write = __gmpz_limbs_write(x : MPZ*, n : MpSize) : MpLimb*
+  fun limbs_finish = __gmpz_limbs_finish(x : MPZ*, s : MpSize)
 
   # MPQ
   struct MPQ